[Gold IV] Title: 감시, Time: 24 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Gold/15683. 감시/감시.cc
@@ -15,7 +15,7 @@ void watch(int cctvDir, int curX, int curY) {
     watch(cctvDir, nX, nY);
 }
 
-void dfs(size_t num) {
+void dfs(int num) {
     if (num == cctv.size()) { // cctv 다 돌았을 경우
         int cnt = 0;
         for (int i = 0; i < N; ++i)
